Add files via upload
bot.py - a script that takes a dictionary of cities and uses it for the game. The user can play with the bot in the cities. The bot processes all possible messages.
cities.txt - list of all cities
cities.py -  converts a list of all cities into a convenient dictionary for quick search